Erased the Command Class. Everything compile and works as before
