--- format_message
@@ -1,5 +1,0 @@
-:prefix(optionnel) COMMAND PARAMETERS
-
-Split by spaces ? 
-: indique 
-15 PARAMETERS MAX
