+++ includes/Irc.hpp
@@ -17,13 +17,9 @@
 
 # include "Macro.hpp"
 # include "Colors.hpp"
+# include "Commands.hpp"
 // #include "Server.hpp" // atttention toi on te surveille
 
-struct cmd_struct
-{
-	std::string	prefix;
-	std::string	name;
-	std::string	message;
-};
+int	parseCommand(std::string cmd_line, cmd_struct &cmd_infos);
 
 #endif
