--- srcs/Commands.cpp
@@ -1,65 +1,0 @@
-#include "Commands.hpp"
-
-Commands::Commands(/* args */) {}
-
-Commands::~Commands() {}
-
-// exple de ligne : @id=234AB :dan!d@localhost PRIVMSG #chan :Hey what's up!
-// autre exple : /ping localhost devient : PRIVMSG localhost :PING 1671195504 289728
-int	Commands::parseCommand(std::string cmd_line, cmd_struct &cmd_infos)
-{
-	// COMMAND
-	for (std::string::size_type i = 0; i < cmd_line.size(); i++)
-	{
-		if (isupper(cmd_line[i]))
-			cmd_infos.name.push_back(cmd_line[i]);
-		if (islower(cmd_line[i]) && cmd_infos.name.empty() == false)
-			break;
-	}
-	
-	// PREFIX
-	size_t prefix_length = cmd_line.find(cmd_infos.name, 0);
-	cmd_infos.prefix.assign(cmd_line, 0, prefix_length);
-	
-	// MESSAGE
-	size_t msg_beginning = cmd_line.find(cmd_infos.name, 0) + cmd_infos.name.length();
-	cmd_infos.message = cmd_line.substr(msg_beginning, std::string::npos);
-
-	// DEBUG
-	std::cout << "Command : " << RED << cmd_infos.name << RESET << std::endl;
-	std::cout << "Prefix : " << BLUE << cmd_infos.prefix << RESET << std::endl;
-	std::cout << "Message : " << GREEN << cmd_infos.message << RESET << std::endl;
-
-	return (SUCCESS);
-}
-
-// /**
-//  * @brief The PING command is sent to check the other side of the connection is still
-//  * 		connected, and/or to check for connection latency.
-//  * 
-//  * 		When receiving a PING message, clients or servers must reply to it with a
-//  * 		PONG message with the same <token> value. 
-//  * 
-//  * 		Numeric replies:
-//  * 		461 - ERR_NEEDMOREPARAMS -  "<client> <command> :Not enough parameters" > DEALT BY IRSSI
-//  * 		409 - ERR_NOORIGIN - "<client> :No origin specified" (when <token> is empty) > DEALT BY IRSSI
-//  * 
-//  * 		IRSSI :
-//  * 		With this reference client, sending a PING to a server sends a CTCP request
-//  * 		to a nickname or a channel. Using the char '*' pings every user in a channel.
-//  * 		Syntax : PING [<nick> | <channel> | *]
-//  * 
-//  * @return SUCCESS (0) or FAILURE (1) 
-//  */
-// int	Commands::ping(int const client_fd, cmd_struct &cmd)
-// {
-// 	// checker la string (est-ce que l'identifier existe (registered or online?), est-ce que le channel existe ?)
-// 	std::cout << client_fd << " is the client_fd. Message de la cmd : " << YELLOW << cmd.message << RESET << std::endl;
-	
-// 	// renvoyer un PONG avec le même TOKEN
-// 	std::string pong_reply = "PONG" + cmd.message + "\r\n";
-// 	send(client_fd, pong_reply.c_str(), pong_reply.size(), 0);
-	
-// 	return (SUCCESS);
-// }
-
